sis. parannuksia kuten: game.price on float eikä Int, pelin editointi on yksittäin, ja toimii djangon formeilla, GameForm sisältää css-luokan. Developer-pelivalikko on vielä kesken (my_games)

+++ website/models.py
@@ -20,7 +20,7 @@ class Game(models.Model):
     name = models.CharField(max_length=255)
     url = models.URLField(blank=False)
     description = models.CharField(max_length=255)
-    price = models.PositiveIntegerField()
+    price = models.FloatField()
     owner = models.ForeignKey(User, on_delete=models.CASCADE)
     def __str__(self):
         return self.name
